+++ src/PointsSumPyramid.cpp
@@ -45,7 +45,7 @@ int PointsSumPyramid::GetOffsetOfLayerN(int n) {
     return offset;
 }
 
-void PointsSumPyramid::PreComputeSum(const Eigen::MatrixXd &cloud_array) {
+void PointsSumPyramid::PreComputeSum(const Eigen::MatrixXf &cloud_array) {
     int offset_this_layer;
     int num_minimum_patches = m_num_elements_each_layer[m_num_layers - 1];
     int num_patch_points = cloud_array.rows() / num_minimum_patches;
@@ -73,8 +73,8 @@ void PointsSumPyramid::PreComputeSum(const Eigen::MatrixXd &cloud_array) {
     }
 }
 
-void PointsSumPyramid::PreComputeSumOfArray(const Eigen::MatrixXd &cloud_array, PointsSum &points_sum_result) {
-    Eigen::MatrixXd X_matrix, Y_matrix, Z_matrix;
+void PointsSumPyramid::PreComputeSumOfArray(const Eigen::MatrixXf &cloud_array, PointsSum &points_sum_result) {
+    Eigen::MatrixXf X_matrix, Y_matrix, Z_matrix;
     X_matrix = cloud_array.col(0);
     Y_matrix = cloud_array.col(1);
     Z_matrix = cloud_array.col(2);
@@ -94,7 +94,7 @@ void PointsSumPyramid::PreComputeSumOfArray(const Eigen::MatrixXd &cloud_array, 
 }
 
 
-int PointsSumPyramid::CountJumps(const Eigen::MatrixXd &Z_matrix, float max_diff) const {
+int PointsSumPyramid::CountJumps(const Eigen::MatrixXf &Z_matrix, float max_diff) const {
     // Check for discontinuities using cross search
     int jumps_count = 0;
     int num_pts_per_cell = Z_matrix.rows();
